feature: change worktree on all buffers
Change update_current_buffer function to update all currently open buffers and de-list buffers from the previous worktree.

Buffers from files that don't exist on another worktree will only be de-listed.

Buffers unrelated to the worktree (temporary or external files) should ignored

The focused buffer file will be kept focused on the new worktree, unless if it doesn't exist, in which case it will select the last one.

In the future, it would also be possible to remember closed files that don't exist for when you switch back, that's not addressed by this PR

Also fixes #84